Remove references to protostore 2.6.*.
While proto2 is still officially supported its now superceded by 3+
on most platforms and is included in mysql libraries making running
both cumbersome.

The bug in the storage 'is set' semantics for vector clock which
was caught some time ago is elided through moving to 1-indexed
vector clocks.

The current InMemoryStore will be refactored to explore better
performance tradeoffs as most operations are currently O(N)
which is sufficient for testing with small N but not for
caching applications.